% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_ccsig.R
\name{df_ccsig}
\alias{df_ccsig}
\title{Smooth the signal and remove the hooks}
\usage{
df_ccsig(
  sig_df,
  span1 = 400,
  span2 = 40,
  breaks_q = c(0, 0.025, 0.05, 0.95, 0.975, 1),
  outlier_cycles = 1,
  ifplot = FALSE
)
}
\arguments{
\item{sig_df}{A signal data frame with two columns: \code{x} and \code{sig}.}

\item{span1}{A positive parameter to control the degree of smoothing in bullterxtrctr::cc_get_signature.}

\item{span2}{A positive parameter to control the degree of smoothing in bullterxtrctr::cc_get_signature.}

\item{breaks_q}{A vector of quantiles to split the \code{x} values into five groups for t-test of large hooks.}

\item{outlier_cycles}{a positive integer for the number of times we check for outliers along the boundaries.}

\item{ifplot}{A Boolean flag indicating whether to save ggplot lists in the output attributes.}
}
\value{
A data frame with two columns:
\itemize{
\item x: The \code{x} values from the \code{x3p} object.
\item sig: The extracted signal.
}
}
\description{
This function smooths the signal and removes the hooks by running a t-test in the boundaries to see if there are large 'hooks'.
}
\examples{
x3p <- x3p_subsamples[[2]]
insidepoly_df <- x3p_insidepoly_df(x3p, mask_col = "#FF0000", concavity = 1.5, b = 1)
x3p_inner_nomiss_res <- df_rmtrend_x3p(insidepoly_df)
x3p_inner_impute <- x3p_impute(x3p_inner_nomiss_res,
  ifout = FALSE, ifsave = FALSE, dir_name = NULL, ifplot = FALSE
)
x3p_bin_rotate <- x3p_vertical(x3p_inner_impute, min_score_cut = 0.1, ifplot = FALSE)
raw_sig_df <- x3p_raw_sig_df(x3p_bin_rotate, ifplot = FALSE)

raw_ccsig_df <- df_ccsig(raw_sig_df, ifplot = TRUE)

attr(raw_ccsig_df, "sig_df_plot")

}
