% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_rmtrend_x3p.R
\name{df_rmtrend_x3p}
\alias{df_rmtrend_x3p}
\title{Removing missing values and a quadratic trend}
\usage{
df_rmtrend_x3p(insidepoly_df)
}
\arguments{
\item{insidepoly_df}{A data frame representing the inner polygon. This is typically obtained from the \code{wire::x3p_insidepoly_df} function.}
}
\value{
An \code{x3p} object that contains the residuals after the removal of the quadratic trend.
}
\description{
This function takes an \code{x3p} object, specifically the inner polygon data frame, and performs two main operations:
\enumerate{
\item Removes missing values.
\item Detrends the data by removing a quadratic trend.
}
}
\examples{
x3p <- x3p_subsamples[[1]]
insidepoly_df <- x3p_insidepoly_df(x3p, mask_col = "#FF0000", concavity = 1.5, b = 1)

x3p_inner_nomiss_res <- df_rmtrend_x3p(insidepoly_df)
x3p_inner_nomiss_res
if (interactive()) {
  x3p_image_autosize(x3p_inner_nomiss_res)
}

}
