% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sig_MSE.R
\name{get_sig_MSE}
\alias{get_sig_MSE}
\title{Extract signal with transformed x3p object by minimizing MSE}
\usage{
get_sig_MSE(x3p, method = "median", ifplot = FALSE, delta = -5:5)
}
\arguments{
\item{x3p}{x3p object}

\item{method}{choice of \code{median} or \code{mean} when computing the summary statistics}

\item{ifplot}{whether graphs are displayed}

\item{delta}{shifting range when minimizing MSE}
}
\description{
Extract signal by computing summary statistics of values along y for each x with transformed x3p object by minimizing MSE.
}
