% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sig_raw.R
\name{get_sig_raw}
\alias{get_sig_raw}
\title{Extract signal with raw x3p object}
\usage{
get_sig_raw(x3p, method = "median", ifplot = FALSE)
}
\arguments{
\item{x3p}{x3p object}

\item{method}{choice of \code{median} or \code{mean} when computing the summary statistics}

\item{ifplot}{whether graphs are displayed}
}
\description{
Extract signal by computing summary statistics of values along y for each x with raw x3p object.
}
