% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sigalign.R
\name{get_sigalign}
\alias{get_sigalign}
\title{Align signals}
\usage{
get_sigalign(
  sig1,
  sig2,
  min.overlap = round(0.75 * min(length(sig1), length(sig2))),
  ifplot = FALSE,
  name1 = "Cut1",
  name2 = "Cut2",
  legendname = "Signal",
  titlename = NULL
)
}
\arguments{
\item{sig1}{first signal vector}

\item{sig2}{second signal vector}

\item{min.overlap}{additional parameter passed on to \code{bulletxtrctr::get_ccf}}

\item{ifplot}{whether graphs are displayed}

\item{name1}{name for the first cut}

\item{name2}{name for the second cut}

\item{legendname}{legend name}

\item{titlename}{title name}
}
\description{
Align signals with plot.
}
