% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_x3p_inner_impute.R
\name{get_x3p_inner_impute}
\alias{get_x3p_inner_impute}
\title{Impute the inner polygon}
\usage{
get_x3p_inner_impute(x3p, mask_col = "#FF0000", concavity = 1.5)
}
\arguments{
\item{x3p}{x3p object}

\item{mask_col}{colour for the polygon}

\item{concavity}{strictly positive value used in \code{concaveman::concaveman}}
}
\description{
Obtained x3p object after imputing the inner polygon.
}
