% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_x3p_inner_nomiss_res.R
\name{get_x3p_inner_nomiss_res}
\alias{get_x3p_inner_nomiss_res}
\title{Removing missing values and the quadratic trend from inner polygon}
\usage{
get_x3p_inner_nomiss_res(x3p, mask_col = "#FF0000", concavity = 1.5)
}
\arguments{
\item{x3p}{x3p object}

\item{mask_col}{colour for the polygon}

\item{concavity}{strictly positive value used in 'concaveman::concaveman'}
}
\description{
Obtained x3p object after removing missing values and the quadratic trend from the inner polygon.
}
