% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_x3p_rotate_angle_MLE.R
\name{get_x3p_rotate_angle_MLE}
\alias{get_x3p_rotate_angle_MLE}
\title{Compute the rotation angle using MLE}
\usage{
get_x3p_rotate_angle_MLE(
  x3p,
  ntheta = 720,
  min_score_cut = 2,
  ifplot = FALSE,
  loess_span = 0.2
)
}
\arguments{
\item{x3p}{x3p object}

\item{ntheta}{number of bins along theta used in \code{imager::hough_line}}

\item{min_score_cut}{the tuning parameter for minimum scores required in hough transformation}

\item{ifplot}{whether graphs are displayed}

\item{loess_span}{the parameter which controls the degree of smoothing}
}
\description{
Compute the rotation angle using MLE with hough transformation.
}
