% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_x3p_rotate_angle_quantile.R
\name{get_x3p_rotate_angle_quantile}
\alias{get_x3p_rotate_angle_quantile}
\title{Compute the rotation angle using quantile}
\usage{
get_x3p_rotate_angle_quantile(
  x3p,
  ntheta = 720,
  min_score_cut = 2,
  ifplot = FALSE
)
}
\arguments{
\item{x3p}{x3p object}

\item{ntheta}{number of bins along theta used in \code{imager::hough_line}}

\item{min_score_cut}{the tuning parameter for minimum scores required in hough transformation}

\item{ifplot}{whether graphs are displayed}
}
\description{
Compute the rotation angle using quantile with hough transformation.
}
