% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inside_polygon.R
\name{inside_polygon}
\alias{inside_polygon}
\title{Identify the inner polygon of a point set}
\usage{
inside_polygon(x, y, concavity, center = NULL)
}
\arguments{
\item{x}{A numeric vector representing the \code{x} coordinates of the points.}

\item{y}{A numeric vector representing the \code{y} coordinates of the points.}

\item{concavity}{A strictly positive parameter that influences the shape of the inner polygon. Smaller values result in a shape that closely follows the inner boundary, while larger values create a shape that focuses more on the central area.}

\item{center}{A numeric vector of length 2, representing the \verb{(x,y)} coordinates of the center point for the mirroring process. If \code{NULL}, the mid-ranges of \code{x} and \code{y} are used.}
}
\value{
A data frame containing \code{x} and \code{y} coordinates that describe the inner polygon. The \code{id} variable indicates the order of the points.
}
\description{
This function identifies the inner polygon of a given set of points. It uses a method of concentric mirroring around a center point, applies an alpha hull to the mirrored shape, and then mirrors the result back.
}
\examples{
x3p <- x3p_subsamples[[1]]
bounds <- x3p_boundary_points(x3p, 2)
polygon <- inside_polygon(bounds$x, bounds$y, 1)

library(ggplot2)
library(dplyr)
bounds \%>\%
  ggplot(aes(x = x, y = y)) +
  geom_point() +
  geom_polygon(data = polygon)

}
