% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_align_sigs_list.R
\name{vec_align_sigs_list}
\alias{vec_align_sigs_list}
\title{Align signals}
\usage{
vec_align_sigs_list(
  sig1,
  sig2,
  min.overlap = round(0.75 * min(length(sig1), length(sig2))),
  ifplot = FALSE,
  name1 = "Cut1",
  name2 = "Cut2",
  legendname = "Signal",
  titlename = NULL
)
}
\arguments{
\item{sig1}{first signal vector}

\item{sig2}{second signal vector}

\item{min.overlap}{additional parameter passed on to \code{bulletxtrctr::get_ccf}}

\item{ifplot}{whether graphs are displayed}

\item{name1}{name for the first cut}

\item{name2}{name for the second cut}

\item{legendname}{legend name}

\item{titlename}{title name}
}
\value{
list of aligned signals named \code{ccf}, \code{lag} and \code{lands} followed the output format of \code{bulletxtrctr::sig_align}
}
\description{
Align signals with plot.
}
