% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_boundary_points.R
\name{x3p_boundary_points}
\alias{x3p_boundary_points}
\title{Extract boundary points from a 3D topographic scan}
\usage{
x3p_boundary_points(x3p, sample)
}
\arguments{
\item{x3p}{An \code{x3p} object representing a topographic scan.}

\item{sample}{A positive integer that defines the sampling rate for the \code{x} and \code{y} directions. Every \code{sample} value will be included in the output. Higher values lead to faster computation but a less precise boundary representation.}
}
\value{
A data frame containing the boundary points. Each row represents a point on the boundary, with \code{x} and \code{y} columns indicating its coordinates.
}
\description{
This function identifies and extracts the boundary points from a 3D topographic scan in \code{x3p} format.
}
\examples{
x3p <- x3p_subsamples[[1]]
bounds <- x3p_boundary_points(x3p, 2)

library(ggplot2)
library(dplyr)
bounds \%>\%
  ggplot(aes(x = x, y = y)) +
  geom_point() +
  theme_bw()

}
