% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_image_autosize.R
\name{x3p_image_autosize}
\alias{x3p_image_autosize}
\title{Display an \code{x3p} object as an auto-sized image}
\usage{
x3p_image_autosize(x3p, ifhtml = FALSE, zoom = 0.6, ...)
}
\arguments{
\item{x3p}{An \code{x3p} object representing a topographic scan.}

\item{ifhtml}{A Boolean flag indicating whether the image will be embedded in HTML.}

\item{zoom}{A numeric value indicating the zoom level for the image display.}

\item{...}{Additional parameters for \code{x3ptools::x3p_image}, excluding \code{size} and \code{zoom}.}
}
\value{
An \code{rgl} plot. If \code{ifhtml = TRUE}, a list of HTML tags is returned.
}
\description{
This function displays an \code{x3p} object as an image with automatically adjusted size based on the dimensions of the \code{x3p} object.
}
\examples{
x3p <- x3p_subsamples[[1]]
if (interactive()) {
  x3p_image_autosize(x3p)
}

}
