% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_impute.R
\name{x3p_impute}
\alias{x3p_impute}
\title{Impute the inner polygon}
\usage{
x3p_impute(x3p, ifsave = FALSE, dir_name = NULL, ifplot = FALSE)
}
\arguments{
\item{x3p}{x3p object}

\item{ifsave}{whether the imputation procedure gif is going to be saved}

\item{dir_name}{required when \code{ifsave} is \code{TRUE}}

\item{ifplot}{whether graphs are displayed}
}
\value{
x3p object after imputation
}
\description{
Obtained x3p object after imputing the inner polygon.
}
\examples{
x3p <- x3p_subsamples[[1]]
insidepoly_df <- x3p_insidepoly_df(x3p, mask_col = "#FF0000", concavity = 1.5, b = 1)
x3p_inner_nomiss_res <- df_rmtrend_x3p(insidepoly_df)

x3p_inner_impute <- x3p_impute(x3p_inner_nomiss_res, ifsave = FALSE, dir_name = NULL, ifplot = TRUE)
x3p_inner_impute
if (interactive()) {
  x3p_image_autosize(x3p_inner_impute)
}

}
