% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_quantile_angle_vec.R
\name{x3p_quantile_angle_vec}
\alias{x3p_quantile_angle_vec}
\title{Calculate rotation angle using quantile method}
\usage{
x3p_quantile_angle_vec(x3p, ntheta = 720, min_score_cut = 0.1, ifplot = FALSE)
}
\arguments{
\item{x3p}{An \code{x3p} object representing a topographic scan.}

\item{ntheta}{An integer representing the number of bins along the theta axis used in \code{imager::hough_line}.}

\item{min_score_cut}{A numeric value representing the minimum score required in the Hough transformation.}

\item{ifplot}{A Boolean flag indicating whether to display graphs.}
}
\value{
A vector of rotation angles computed by the quantile procedure.
}
\description{
This function calculates the rotation angle of an \code{x3p} object using the quantile method with Hough transformation.
}
\examples{
x3p <- x3p_subsamples[[1]]
insidepoly_df <- x3p_insidepoly_df(x3p, mask_col = "#FF0000", concavity = 1.5, b = 1)
x3p_inner_nomiss_res <- df_rmtrend_x3p(insidepoly_df)
x3p_inner_impute <- x3p_impute(x3p_inner_nomiss_res,
ifout = FALSE, ifsave = FALSE, dir_name = NULL, ifplot = FALSE)

x3p_quantile_angle_vec(x3p_inner_impute, min_score_cut = 0.1, ifplot = TRUE) \%>\%
    str()

}
