% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_raw_sig_vec.R
\name{x3p_raw_sig_vec}
\alias{x3p_raw_sig_vec}
\title{Extract signal with raw x3p object}
\usage{
x3p_raw_sig_vec(x3p, method = "median", ifplot = FALSE)
}
\arguments{
\item{x3p}{x3p object}

\item{method}{choice of \code{median} or \code{mean} when computing the summary statistics}

\item{ifplot}{whether graphs are displayed}
}
\value{
vector of raw signal extracted
}
\description{
Extract signal by computing summary statistics of values along y for each x with raw x3p object.
}
