% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_shift.R
\name{x3p_shift}
\alias{x3p_shift}
\title{Minimize MSE by shifting striations}
\usage{
x3p_shift(x3p, ifplot = FALSE, delta = -5:5, delta_q_range = c(0, 1))
}
\arguments{
\item{x3p}{An \code{x3p} object representing a topographic scan.}

\item{ifplot}{A Boolean flag indicating whether to display graphs.}

\item{delta}{A numeric vector representing the shifting range for minimizing MSE.}

\item{delta_q_range}{A numeric vector of length 2, representing the lower and upper bounds for the quantile taken.}
}
\value{
An \code{x3p} object after the transformation.
}
\description{
This function shifts the striations on an \code{x3p} object to minimize the Mean Squared Error (MSE).
}
\examples{
x3p <- x3p_subsamples[[2]]

insidepoly_df <- x3p_insidepoly_df(x3p, mask_col = "#FF0000", concavity = 1.5, b = 1)
x3p_inner_nomiss_res <- df_rmtrend_x3p(insidepoly_df)
x3p_inner_impute <- x3p_impute(x3p_inner_nomiss_res,
  ifout = FALSE, ifsave = FALSE, dir_name = NULL, ifplot = FALSE
)
x3p_bin_rotate <- x3p_vertical(x3p_inner_impute, min_score_cut = 0.1, ifplot = FALSE)

if (interactive()) {
  x3p_approx <- x3p_shift(x3p_bin_rotate, ifplot = TRUE)

  attr(x3p_approx, "x3p_before_shift_plot")
  attr(x3p_approx, "x3p_after_shift_plot")
  attr(x3p_approx, "MSE_plot")
}

}
