% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_shift_midlag.R
\name{x3p_shift_midlag}
\alias{x3p_shift_midlag}
\title{Shift striations by minimizing MSE}
\usage{
x3p_shift_midlag(x3p, ifplot = FALSE, delta = -5:5)
}
\arguments{
\item{x3p}{\code{x3p} object}

\item{ifplot}{whether graphs are displayed}

\item{delta}{shifting range when minimizing MSE}
}
\value{
\code{x3p} object after transformation
}
\description{
Shift striations on \code{x3p} object by minimizing MSE.
}
\examples{
x3p <- x3p_subsamples[[2]]

insidepoly_df <- x3p_insidepoly_df(x3p, mask_col = "#FF0000", concavity = 1.5, b = 1)
x3p_inner_nomiss_res <- df_rmtrend_x3p(insidepoly_df)
x3p_inner_impute <- x3p_impute(x3p_inner_nomiss_res,
  ifout = FALSE, ifsave = FALSE, dir_name = NULL, ifplot = FALSE
)
x3p_bin_rotate <- x3p_vertical(x3p_inner_impute, min_score_cut = 0.1, ifplot = FALSE)

if (interactive()) {
  x3p_shift_midlag(x3p_bin_rotate, ifplot = TRUE)
}

}
