% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{x3p_subsamples}
\alias{x3p_subsamples}
\title{Subsampled wire cuts data}
\format{
list
A list of x3p objects
}
\usage{
x3p_subsamples
}
\description{
A list of wire cuts data set by subsampling factor 10 of the original wire cuts x3p.
}
\keyword{datasets}
