% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_surface_polygon.R
\name{x3p_surface_polygon}
\alias{x3p_surface_polygon}
\title{Add a polygon mask of the scan shape}
\usage{
x3p_surface_polygon(
  x3p,
  colour = "red",
  sample = 10,
  center = NULL,
  concavity = 1.5
)
}
\arguments{
\item{x3p}{An \code{x3p} object representing a topographic scan.}

\item{colour}{A string specifying the colour of the polygon mask.}

\item{sample}{A positive integer indicating the downsampling rate used to calculate the boundary points of the polygon. Higher values result in faster computation but less accurate polygons.}

\item{center}{A numeric vector representing the center point of the scan. If \code{NULL}, the center point is derived from the boundary points.}

\item{concavity}{A positive number used in the \code{concaveman::concaveman} function to determine the concavity of the polygon.}
}
\value{
An \code{x3p} object with an added polygon mask.
}
\description{
This function adds a polygon mask to an \code{x3p} scan. The polygon represents the shape of the scan.
}
\examples{
if (interactive()) {
  x3p <- x3p_subsamples[[1]]
  x3p <- x3p \%>\% x3p_surface_polygon(sample=1)
  x3p_image(x3p, size = dim(x3p$surface.matrix), zoom=.6)
}

}
