% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_rmtrend_x3p.R
\name{df_rmtrend_x3p}
\alias{df_rmtrend_x3p}
\title{Removing missing values and the quadratic trend from inner polygon}
\usage{
df_rmtrend_x3p(x3p, mask_col = "#FF0000", concavity = 1.5)
}
\arguments{
\item{x3p}{x3p object}

\item{mask_col}{colour for the polygon}

\item{concavity}{strictly positive value used in \code{concaveman::concaveman}}
}
\description{
Obtained x3p object after removing missing values and the quadratic trend from the inner polygon.
}
