% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_rmtrend_x3p.R
\name{df_rmtrend_x3p}
\alias{df_rmtrend_x3p}
\title{Removing missing values and the quadratic trend from inner polygon}
\usage{
df_rmtrend_x3p(insidepoly_df)
}
\arguments{
\item{insidepoly_df}{data frame of inside polygon obtained from \code{wire::x3p_insidepoly_df}}
}
\value{
\code{x3p} object of residuals after removing trend
}
\description{
Obtained \code{x3p} object after removing missing values and the quadratic trend from the inner polygon.
}
\examples{
x3p <- x3p_subsamples[[1]]
insidepoly_df <- x3p_insidepoly_df(x3p, mask_col = "#FF0000", concavity = 1.5, b = 1)

x3p_inner_nomiss_res <- df_rmtrend_x3p(insidepoly_df)
x3p_inner_nomiss_res
if (interactive()) {
  x3p_image_autosize(x3p_inner_nomiss_res)
}

}
