% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_x3p_inner_df.R
\name{get_x3p_inner_df}
\alias{get_x3p_inner_df}
\title{Create the summary data frame for the inner part of an x3p object}
\usage{
get_x3p_inner_df(x3p, mask_col = "#FF0000", concavity = 1.5)
}
\arguments{
\item{x3p}{x3p object}

\item{mask_col}{colour for the polygon}

\item{concavity}{strictly positive value used in 'concaveman::concaveman'}
}
\description{
Create the summary data frame for the inner part of an x3p object.
}
