% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inside_polygon.R
\name{inside_polygon}
\alias{inside_polygon}
\title{Identify the inside polygon of a set of points}
\usage{
inside_polygon(x, y, concavity, center = NULL)
}
\arguments{
\item{x}{numeric vector of \code{x} positions}

\item{y}{numeric vector of \code{y} positions}

\item{concavity}{strictly positive parameter. Smaller values approach the
inside of the boundary
more closely. Large values concentrate on the inner center of the area}

\item{center}{numeric vector of length 2, consisting of \verb{(x,y)} coordinate for
centering \code{x} and \code{y} values.
If center is \code{NULL}, the half ranges of \code{x} and \code{y} are used.}
}
\value{
data frame of \code{x} and \code{y} positions describing the inside of the area
described by the input \code{x} and \code{y}, variable id describes the order of the points
}
\description{
Uses concentric mirroring onto a radius around the center point, applies an
alpha hull to that shape, and mirrors the result back.
}
\examples{
x3p <- x3p_subsamples[[1]]
bounds <- x3p_boundary_points(x3p, 2)
polygon <- inside_polygon(bounds$x, bounds$y, 1)

library(ggplot2)
library(dplyr)
bounds \%>\%
ggplot(aes(x = x, y = y)) + geom_point() +
  geom_polygon(data = polygon)

}
