% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_align_sigs_list.R
\name{vec_align_sigs_list}
\alias{vec_align_sigs_list}
\title{Align two signal vectors}
\usage{
vec_align_sigs_list(
  sig1,
  sig2,
  min.overlap = NULL,
  ifplot = FALSE,
  name1 = "Cut1",
  name2 = "Cut2",
  legendname = "Signal",
  titlename = NULL
)
}
\arguments{
\item{sig1}{The first numeric signal vector.}

\item{sig2}{The second numeric signal vector.}

\item{min.overlap}{An optional parameter passed to \code{bulletxtrctr::get_ccf} to specify the minimum overlap between signals.}

\item{ifplot}{A Boolean flag indicating whether to plot the aligned signals.}

\item{name1}{A string to label the first signal in the plot.}

\item{name2}{A string to label the second signal in the plot.}

\item{legendname}{A string to label the legend in the plot.}

\item{titlename}{A string to set the title of the plot.}
}
\value{
A list containing the cross-correlation function (\code{ccf}), the lag (\code{lag}), and the landmarks (\code{lands}) of the aligned signals. This follows the output format of \code{bulletxtrctr::sig_align}.
}
\description{
This function aligns two numeric signal vectors. It also provides an option to visualize the aligned signals.
}
\examples{
x3p <- x3p_subsamples[[1]]
insidepoly_df <- x3p_insidepoly_df(x3p, mask_col = "#FF0000", concavity = 1.5, b = 1)
x3p_inner_nomiss_res <- df_rmtrend_x3p(insidepoly_df)
x3p_inner_impute <- x3p_impute(x3p_inner_nomiss_res,
  ifout = FALSE, ifsave = FALSE, dir_name = NULL, ifplot = FALSE
)
x3p_bin_rotate <- x3p_vertical(x3p_inner_impute, min_score_cut = 0.1)

vec_align_sigs_list(x3p_raw_sig_vec(x3p_bin_rotate)$sig, x3p_shift_sig_vec(x3p_bin_rotate)$sig,
  ifplot = TRUE
) \%>\%
  str()

}
