% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_image_autosize.R
\name{x3p_image_autosize}
\alias{x3p_image_autosize}
\title{Plot x3p object as an image with auto-adjusted size}
\usage{
x3p_image_autosize(x3p, zoom = 0.6, ...)
}
\arguments{
\item{x3p}{x3p object}

\item{zoom}{numeric value indicating the amount of zoom}

\item{...}{other parameter values except size and zoom used in \code{x3ptools::x3p_image}}
}
\description{
Plot x3p object as an image with auto-adjusted size based on \code{x3ptools::x3p_image}.
}
