% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_impute.R
\name{x3p_impute}
\alias{x3p_impute}
\title{Impute the inner polygon}
\usage{
x3p_impute(
  x3p,
  x3p_mask,
  mask_col = "#FF0000",
  concavity = 1.5,
  ifsave = FALSE,
  dir_name = NULL,
  ifplot = FALSE
)
}
\arguments{
\item{x3p}{x3p object}

\item{x3p_mask}{x3p object for mask}

\item{mask_col}{colour for the polygon}

\item{concavity}{strictly positive value used in \code{concaveman::concaveman}}

\item{ifsave}{whether the imputation procedure gif is going to be saved}

\item{dir_name}{required when \code{ifsave} is \code{TRUE}}

\item{ifplot}{whether graphs are displayed}
}
\value{
x3p object after imputation
}
\description{
Obtained x3p object after imputing the inner polygon.
}
\examples{
x3p <- x3p_subsamples[[1]]
mask_col <- "#FF0000"
concavity <- 1.5

insidepoly_df <- x3p_insidepoly_df(x3p, mask_col = mask_col, concavity = concavity)
x3p_inner_nomiss_res <- df_rmtrend_x3p(insidepoly_df)

x3p_inner_impute <- x3p_impute(x3p_inner_nomiss_res, x3p, mask_col = mask_col,
concavity = concavity, ifsave = FALSE, dir_name = NULL, ifplot = TRUE)
x3p_inner_impute

if (interactive()) {
  x3p_image_autosize(x3p_inner_impute)
}

}
