% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_impute.R
\name{x3p_impute}
\alias{x3p_impute}
\title{Impute the inner polygon}
\usage{
x3p_impute(
  x3p,
  mask_col = "#FF0000",
  concavity = 1.5,
  ifsave = FALSE,
  dir_name = NULL,
  ifplot = FALSE
)
}
\arguments{
\item{x3p}{x3p object}

\item{mask_col}{colour for the polygon}

\item{concavity}{strictly positive value used in \code{concaveman::concaveman}}

\item{ifsave}{whether the imputation procedure gif is going to be saved}

\item{dir_name}{required when \code{ifsave} is \code{TRUE}}

\item{ifplot}{whether graphs are displayed}
}
\value{
x3p obeject after imputation
}
\description{
Obtained x3p object after imputing the inner polygon.
}
