% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_insidepoly_df.R
\name{x3p_insidepoly_df}
\alias{x3p_insidepoly_df}
\title{Summary data frame for the inner polygon}
\usage{
x3p_insidepoly_df(x3p, mask_col = "#FF0000", concavity = 1.5, b = 10)
}
\arguments{
\item{x3p}{x3p object}

\item{mask_col}{colour for the polygon}

\item{concavity}{strictly positive value used in \code{concaveman::concaveman}}

\item{b}{positive integer value, block size, used in \code{x3ptools::x3p_average}}
}
\value{
data frame of inside polygon
\itemize{
\item{x}{x value from input x3p object}
\item{y}{y value from input x3p object}
\item{value}{height value from input x3p object}
\item{mask}{mask value from input x3p object}
\item{n_neighbor_val_miss}{number of missing immediate neighbor, self included}
\item{sd_not_miss}{standard deviation for immediate neighbor}
}
}
\description{
Create the summary data frame for the inner polygon of an x3p object.
}
\examples{
x3p <- x3p_subsamples[[1]]

x3p_insidepoly_df(x3p, mask_col = "#FF0000", concavity = 1.5, b = 1) \%>\%
str()

}
