% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_insidepoly_df.R
\name{x3p_insidepoly_df}
\alias{x3p_insidepoly_df}
\title{Generate summary data frame for inner polygon}
\usage{
x3p_insidepoly_df(
  x3p,
  mask_col = "#FF0000",
  concavity = 1.5,
  b = 10,
  ifplot = FALSE
)
}
\arguments{
\item{x3p}{An \code{x3p} object representing a topographic scan.}

\item{mask_col}{A string representing the color to be used for the polygon.}

\item{concavity}{A strictly positive number used in \code{concaveman::concaveman} to influence the shape of the polygon.}

\item{b}{A positive integer representing the block size for \code{x3ptools::x3p_average}.}

\item{ifplot}{A Boolean flag indicating whether to display graphs.}
}
\value{
A data frame summarizing the inner polygon. The data frame includes the following columns:
\itemize{
\item x: The \code{x} coordinates from the \code{x3p} object.
\item y: The \code{y} coordinates from the \code{x3p} object.
\item value: The height values from the \code{x3p} object.
\item mask: The mask values from the \code{x3p} object.
\item n_neighbor_val_miss: The number of immediate neighbors, including the point itself, that are missing.
\item sd_not_miss: The standard deviation of the immediate neighbors that are not missing.
}
}
\description{
This function generates a summary data frame for the inner polygon of an \code{x3p} object.
}
\examples{
x3p <- x3p_subsamples[[1]]

insidepoly_df <- x3p_insidepoly_df(x3p, mask_col = "#FF0000", concavity = 1.5, b = 1, ifplot = TRUE)

attr(insidepoly_df, "x3p_plot")
attr(insidepoly_df, "number_of_missing_immediate_neighbors_plot")
attr(insidepoly_df, "standard_deviation_of_non_missing_immediate_neighbors_plot")
attr(insidepoly_df, "number_of_missing_immediate_neighbors_boxplot")

}
