% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_MLE_angle_vec.R
\name{x3p_MLE_angle_vec}
\alias{x3p_MLE_angle_vec}
\title{Calculate rotation angle using maximum likelihood estimation (MLE)}
\usage{
x3p_MLE_angle_vec(
  x3p,
  ntheta = 720,
  min_score_cut = 0.1,
  ifplot = FALSE,
  loess_span = 0.2
)
}
\arguments{
\item{x3p}{An \code{x3p} object representing a topographic scan.}

\item{ntheta}{The number of bins along the theta axis used in \code{imager::hough_line}.}

\item{min_score_cut}{A tuning parameter that sets the minimum score required for the Hough transformation.}

\item{ifplot}{A Boolean flag indicating whether to display graphs.}

\item{loess_span}{A parameter controlling the degree of smoothing in the LOESS function.}
}
\value{
A vector of rotation angles computed by the MLE procedure.
}
\description{
This function calculates the rotation angle of an \code{x3p} object using maximum likelihood estimation (MLE) with Hough transformation.
}
\examples{
x3p <- x3p_subsamples[[1]]
insidepoly_df <- x3p_insidepoly_df(x3p, mask_col = "#FF0000", concavity = 1.5, b = 1)
x3p_inner_nomiss_res <- df_rmtrend_x3p(insidepoly_df)
x3p_inner_impute <- x3p_impute(x3p_inner_nomiss_res,
  ifout = FALSE, ifsave = FALSE, dir_name = NULL, ifplot = FALSE
)

x3p_bin <- x3p_inner_impute \%>\%
  x3ptools::x3p_bin_stripes(
    direction = "vertical",
    colors = c("#b12819", "#ffffff", "#134D6B"),
    freqs = c(0, 0.3, 0.7, 1)
  )
x3p_bin_red <- x3ptools::x3p_extract(x3p_bin, mask_vals = "#b12819")

x3p_MLE_angle_vec(x3p_bin_red, min_score_cut = 5, ifplot = TRUE) \%>\%
  str()

}
