% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_raw_sig_vec.R
\name{x3p_raw_sig_vec}
\alias{x3p_raw_sig_vec}
\title{Extract signal from raw \code{x3p} object}
\usage{
x3p_raw_sig_vec(x3p, method = "median", ifplot = FALSE)
}
\arguments{
\item{x3p}{An \code{x3p} object representing a topographic scan.}

\item{method}{A string indicating the method for computing summary statistics. Options are \code{median} or \code{mean}.}

\item{ifplot}{A Boolean flag indicating whether to display graphs.}
}
\value{
A data frame with two columns:
\itemize{
\item x: The \code{x} values from the \code{x3p} object.
\item sig: The extracted signal.
}
}
\description{
This function extracts the signal from a raw \code{x3p} object by computing summary statistics of values along the \code{y} axis for each \code{x} value.
}
\examples{
x3p <- x3p_subsamples[[2]]
insidepoly_df <- x3p_insidepoly_df(x3p, mask_col = "#FF0000", concavity = 1.5, b = 1)
x3p_inner_nomiss_res <- df_rmtrend_x3p(insidepoly_df)
x3p_inner_impute <- x3p_impute(x3p_inner_nomiss_res,
  ifout = FALSE, ifsave = FALSE, dir_name = NULL, ifplot = FALSE
)
x3p_bin_rotate <- x3p_vertical(x3p_inner_impute, min_score_cut = 0.1, ifplot = FALSE)

x3p_raw_sig_vec(x3p_bin_rotate, ifplot = TRUE) \%>\%
  str()

}
