% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_shift_sig_vec.R
\name{x3p_shift_sig_vec}
\alias{x3p_shift_sig_vec}
\title{Extract signal with transformed x3p object by minimizing MSE}
\usage{
x3p_shift_sig_vec(x3p, method = "median", ifplot = FALSE, delta = -5:5)
}
\arguments{
\item{x3p}{x3p object}

\item{method}{choice of \code{median} or \code{mean} when computing the summary statistics}

\item{ifplot}{whether graphs are displayed}

\item{delta}{shifting range when minimizing MSE}
}
\value{
vector of shifted signals
}
\description{
Extract signal by computing summary statistics of values along y for each x with transformed x3p object by minimizing MSE.
}
\examples{
x3p <- x3p_subsamples[[2]]

insidepoly_df <- x3p_insidepoly_df(x3p, mask_col = "#FF0000", concavity = 1.5, b = 1)
x3p_inner_nomiss_res <- df_rmtrend_x3p(insidepoly_df)
x3p_inner_impute <- x3p_impute(x3p_inner_nomiss_res, ifsave = FALSE, dir_name = NULL, ifplot = TRUE)
x3p_bin_rotate <- x3p_vertical(x3p_inner_impute, min_score_cut = 0.1)

if (interactive()) {
  x3p_shift_sig_vec(x3p_bin_rotate, ifplot = TRUE) \%>\%
  str()
}

}
