% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{x3p_subsamples}
\alias{x3p_subsamples}
\title{Subsampled wire cuts data}
\format{
list
A list of \code{x3p} objects named with surface scan names
\itemize{
\item T2AW-LM-R2-B32: \code{x3p} object
\item T2CW-LI-R2-B15: \code{x3p} object
}
}
\usage{
x3p_subsamples
}
\description{
A list of wire cuts data set by subsampling factor 10 of the original wire cuts \code{x3p}.
}
\keyword{datasets}
