% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_surface_polygon.R
\name{x3p_surface_polygon}
\alias{x3p_surface_polygon}
\title{Add a polygon of the scan shape as mask}
\usage{
x3p_surface_polygon(
  x3p,
  colour = "red",
  sample = 10,
  center = NULL,
  concavity = 1.5
)
}
\arguments{
\item{x3p}{x3p file}

\item{colour}{colour for the polygon}

\item{sample}{strictly positive integer value specifying the amount of
downsampling to be used (see \code{x3ptools::x3p_sample}) to calculate the boundary points
of the polygon. Larger values of \code{sample} will result in faster response times
but results in less accurate polygons.}

\item{center}{point on the center of the scan. If NULL, a center will be
derived from the boundary points}

\item{concavity}{strictly positive value used in \code{concaveman::concaveman}}
}
\value{
x3p object with mask
}
\description{
Add a polygon of the scan shape as mask
}
\examples{
if (interactive()) {
  x3p <- x3p_subsamples[[1]]
  x3p <- x3p \%>\% x3p_surface_polygon(sample=1)
  x3p_image(x3p, size = dim(x3p$surface.matrix), zoom=.6)
}

}
