% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_vertical.R
\name{x3p_vertical}
\alias{x3p_vertical}
\title{Rotate imputed \code{x3p} object to vertical direction}
\usage{
x3p_vertical(
  x3p_inner_impute,
  freqs = c(0, 0.3, 0.7, 1),
  method = "MLE",
  ntheta = 720,
  min_score_cut = 0.1,
  ifplot = FALSE,
  loess_span = 0.2
)
}
\arguments{
\item{x3p_inner_impute}{An \code{x3p} object after imputation.}

\item{freqs}{A numeric vector of length 4, representing color frequency quantiles used in \code{x3ptools::x3p_bin_stripes}.}

\item{method}{A string indicating the method for computing rotation angle. Options are \code{MLE} or \code{quantile}.}

\item{ntheta}{An integer representing the number of bins along the theta axis used in \code{imager::hough_line}.}

\item{min_score_cut}{A numeric value representing the minimum score required in the Hough transformation.}

\item{ifplot}{A Boolean flag indicating whether to display graphs.}

\item{loess_span}{A numeric value controlling the degree of smoothing. This is only available when \code{method = MLE}.}
}
\value{
An \code{x3p} object after rotation with vertical striations.
}
\description{
This function rotates an imputed \code{x3p} object with striations to the vertical direction.
}
\examples{
x3p <- x3p_subsamples[[1]]

insidepoly_df <- x3p_insidepoly_df(x3p, mask_col = "#FF0000", concavity = 1.5, b = 1)
x3p_inner_nomiss_res <- df_rmtrend_x3p(insidepoly_df)
x3p_inner_impute <- x3p_impute(x3p_inner_nomiss_res,
  ifout = FALSE, ifsave = FALSE, dir_name = NULL, ifplot = FALSE
)

x3p_bin_rotate <- x3p_vertical(x3p_inner_impute, min_score_cut = 0.1, ifplot = TRUE)
x3p_bin_rotate
if (interactive()) {
  x3p_image_autosize(x3p_bin_rotate)
}

}
