% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_vertical.R
\name{x3p_vertical}
\alias{x3p_vertical}
\title{Rotate imputed x3p object to vertical direction}
\usage{
x3p_vertical(
  x3p_inner_impute,
  freqs = c(0, 0.3, 0.7, 1),
  method = "MLE",
  ntheta = 720,
  min_score_cut = 0.1,
  ifplot = FALSE,
  loess_span = 0.2
)
}
\arguments{
\item{x3p_inner_impute}{x3p object after imputation}

\item{freqs}{length 4 vector of values corresponding to color frequency (turned into quantiles of the differenced values) used in \code{x3ptools::x3p_bin_stripes}}

\item{method}{choice of \code{MLE} or \code{quantile} when computing rotation angle}

\item{ntheta}{number of bins along theta used in \code{imager::hough_line}}

\item{min_score_cut}{the tuning parameter for minimum scores required in hough transformation}

\item{ifplot}{whether graphs are displayed}

\item{loess_span}{the parameter which controls the degree of smoothing, only available when \code{method = MLE}}
}
\value{
x3p object after rotation with vertical striations
}
\description{
Rotate imputed x3p object with striations to the vertical direction.
}
\examples{
x3p <- x3p_subsamples[[1]]
mask_col <- "#FF0000"
concavity <- 1.5

insidepoly_df <- x3p_insidepoly_df(x3p, mask_col = mask_col, concavity = concavity, b = 1)
x3p_inner_nomiss_res <- df_rmtrend_x3p(insidepoly_df)
x3p_inner_impute <- x3p_impute(x3p_inner_nomiss_res, ifsave = FALSE, dir_name = NULL, ifplot = TRUE)

x3p_bin_rotate <- x3p_vertical(x3p_inner_impute, min_score_cut = 0.1, ifplot = TRUE)
x3p_bin_rotate

if (interactive()) {
  x3p_image_autosize(x3p_bin_rotate)
}

}
